Scan Claude launch options past prompt positionals on resume
Fix the truncation class in AgentLaunchSanitizer.preserveOptions for
Claude: a new claude-only Policy knob (scansOptionsPastPositionals)
skips prompt positionals and keeps scanning instead of ending the scan,
gives unknown options a fail-safe arity heuristic so an unlisted value
flag never truncates or corrupts the tail, and bounds variadic value
consumption at prompt-shaped tokens. Only the first positional can
still match nonRestorableCommands; prompt positionals and
session-binding flags are still dropped, "--" still ends the scan, and
Claude Teams keeps its conservative post-option prompt boundary so
flag-shaped tokens spilled from a --tmux prompt payload are never
promoted to options.

The scanner moves to AgentLaunchPositionalScanning.swift and grokPolicy
moves verbatim to AgentLaunchSanitizerGrokPolicy.swift to stay inside
the Swift file length budget without touching the TSV.

Closes https://github.com/manaflow-ai/cmux/issues/7599
