+++ Packages/macOS/CMUXAgentLaunch/Sources/CMUXAgentLaunch/AgentLaunchSanitizerPrimaryPolicies.swift
@@ -51,6 +51,7 @@ extension AgentLaunchSanitizer {
             "--allowedTools",
             "--allowed-tools",
             "--betas",
+            "--dangerously-load-development-channels",
             "--disallowedTools",
             "--disallowed-tools",
             "--file",
@@ -101,6 +102,7 @@ extension AgentLaunchSanitizer {
             "-p",
             "--no-session-persistence"
         ],
+        scansOptionsPastPositionals: true,
         skipClaudeHookSettings: true
     )
 
@@ -165,80 +167,6 @@ extension AgentLaunchSanitizer {
             "--remote-auth-token-env="
         ],
         resumeSubcommand: "resume"
-    )
-
-    static let grokPolicy = Policy(
-        valueOptions: [
-            "--agent",
-            "--agents",
-            "--allow",
-            "--cwd",
-            "--deny",
-            "--disallowed-tools",
-            "--effort",
-            "--max-turns",
-            "--model",
-            "-m",
-            "--permission-mode",
-            "--reasoning-effort",
-            "--resume",
-            "-r",
-            "--rules",
-            "--sandbox",
-            "--system-prompt-override",
-            "--tools",
-            "--worktree",
-            "-w"
-        ],
-        optionalValueOptions: [
-            "--resume",
-            "-r",
-            "--worktree",
-            "-w"
-        ],
-        nonRestorableCommands: [
-            "agent",
-            "help",
-            "import",
-            "inspect",
-            "leader",
-            "login",
-            "mcp",
-            "memory",
-            "models",
-            "sessions",
-            "setup",
-            "share",
-            "ssh",
-            "trace",
-            "update",
-            "version",
-            "v",
-            "worktree"
-        ],
-        droppedOptions: [
-            "--continue",
-            "-c",
-            "--restore-code",
-            "--resume",
-            "-r",
-            "--worktree",
-            "-w"
-        ],
-        droppedOptionPrefixes: [
-            "--resume=",
-            "-r=",
-            "--worktree=",
-            "-w="
-        ],
-        rejectOptions: [
-            "--best-of-n",
-            "--output-format",
-            "--prompt-file",
-            "--prompt-json",
-            "--single",
-            "-p"
-        ]
     )
 
     static let piPolicy = Policy(
